+++ Source/Examples/ExampleLibrary/Annotations/TileMapAnnotationExamples.cs
@@ -16,6 +16,8 @@ namespace ExampleLibrary
         [Example("TileMapAnnotation (openstreetmap.org)")]
         public static PlotModel TileMapAnnotation2()
         {
+            // See policy document: https://operations.osmfoundation.org/policies/tiles/
+
             var model = new PlotModel { Title = "TileMapAnnotation" };
             model.Axes.Add(new LinearAxis { Position = AxisPosition.Bottom, Minimum = 10.4, Maximum = 10.6, Title = "Longitude" });
             model.Axes.Add(new LinearAxis { Position = AxisPosition.Left, Minimum = 59.88, Maximum = 59.96, Title = "Latitude" });
@@ -25,7 +27,8 @@ namespace ExampleLibrary
                 new TileMapAnnotation
                     {
                         Url = "http://tile.openstreetmap.org/{Z}/{X}/{Y}.png",
-                        CopyrightNotice = "OpenStreetMap"
+                        CopyrightNotice = "OpenStreetMap",
+                        MaxNumberOfDownloads = 2,
                     });
 
             return model;