+++ CHANGELOG.md
@@ -103,6 +103,7 @@ All notable changes to this project will be documented in this file.
 - MinimumPadding incorrect when MaximumPadding is non-zero (#1625)
 - Don't clip zerocrossing axis lines within plot bounds (#1441)
 - Incorrect margins when using Color Axes with AxisPosition.None (#1574)
+- OpenStreetMap example (#1642)
 
 ## [2.0.0] - 2019-10-19
 ### Added